backend/requirements.txt
fastapi
uvicorn
pdfminer.six
docx2txt
chromadb
pydantic
python-multipart
google-generativeai
tiktoken
requests
